[Modify] Navigation Structure
